Remove legacy runner adapter to request checks
