+++ config/config.exs
@@ -123,7 +123,6 @@ config :trento, Trento.Integration.Telemetry, adapter: Trento.Integration.Teleme
 config :trento, :messaging, adapter: Trento.Messaging.Adapters.AMQP
 config :trento, Trento.Integration.Checks, adapter: Trento.Integration.Checks.Wanda
 config :trento, Trento.Messaging.Publisher, adapter: Trento.Messaging.Adapters.AMQP
-config :trento, Trento.Clusters, checks_adapter: Trento.Clusters.Wanda
 
 config :trento, Trento.Messaging.Adapters.AMQP,
   publisher: [
